+++ lib/ESP_WiFiManager_Lite/changelog.md
@@ -45,6 +45,7 @@
 
 1. Massive simplification to reset detector configuration
 2. Using the simplified MultiResetDetector v2.0, makes the DoubleResetDetector library superfluous
+3. BREAKING CHANGE: replaced everywhere "Customs" with "Custom" (spelling error introduced in version 1.2.0). This affects users with configurable **Custom HTML Headers**, including Custom Style and Custom Head Elements.
 
 ### Release v1.10.5